requirements.txt
click>=8.0.0
customtkinter>=5.2.0
markdown-it-py>=3.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-dotenv>=1.0.0
olefile>=0.46
pdfminer.six>=20221105
PyPDF2>=3.0.0